requirements.txt
matplotlib
autopep8
requests
python-dotenv
web3
SQLAlchemy
scipy
psycopg2-binary
